services/agents/curator/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-pubsub
google-cloud-secret-manager
google-generativeai
